requirements.txt
pandas
numpy
psycopg2-binary
requests
torch
transformers
accelerate
bitsandbytes
scikit-learn
tqdm
dotenv